Fix std::less<SAGEExpr> specialization namespace.
When specializing a template from some namespace foo, the specialization
must itself be declared within "namespace foo { ... }".  It's not enough
to just prefix the specialization with "foo::".  For more information
and commentary, see <http://stackoverflow.com/q/25594644/91929>,
<http://stackoverflow.com/q/2282349/91929>,
<http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_defects.html#374>, and
<https://gcc.gnu.org/bugzilla/show_bug.cgi?id=56480>.

Note that it is only necessary to *declare* the specialization in this
manner; it is not required to also *define* the specialization within
namespace curly brackets.  However, one is allowed to do that where
convenient, as I have done here.

This fixes a build failure with g++-4.8.3, and probably with many other
gcc and clang versions as well.