Ignore object files with common suffixes
The ".os" suffix would not typically appear in a standard LLVM build
tree.  However, SCons uses this as its default suffix for object files
built with position-independent code, such as would eventually be linked
into a shared library or plugin.  So ignoring this suffix too can be
useful in case someone uses this code within a larger, SCons-managed
project.