Ignore generated object files in standalone builds
These ignore patterns are useful when this is being built as a
standalone project, not as part of an LLVM build tree.  In that
scenario, generated object files likely appear directly in the source
tree, not in some other directory.  The "*.so" pattern ignores the final
plugin, while "*.os" is for shared object files containing
position-independent code.  (The latter suffix is used by scons when
creating object files that will eventually be linked into shared
libraries.)